A couple of small fixes (#369)
* Fix es_mx lang file

* Fix CrabEntity.Home serialization

* Fix crab home location

The crab home location used the position before it was actually set.
The home location of crabs from crab eggs was always 0 0 0

* Fix NullPointerException crash

Fixes a crash that occurred when spawning a baby crab by using a spawn
egg on a crab. In this case the home is never set and causes an NPE
when it tries to save.